feature(Embedding): added `DiscretizedProgramStaticEmbeddings` (#43)
* feature(Embedding): encapsulate embedding (and reverse) functions and loss into `ProgramStaticEmbeddings`

* fix(Tests): integration tests should run now

* Create black.yaml

* fix(Config): removed unnecessary instruction_embedding_size

* refactor(Embedding): renamed file

* fix(Tests): unit test

* feature(Embedding): added `DiscretizedProgramStaticEmbeddings`

+++ embedding/utils.py
@@ -7,7 +7,11 @@ def quantize(input: float, n: int) -> torch.Tensor:
 
 
 def to_onehot(x: int, n_classes: int) -> torch.Tensor:
-    return torch.nn.functional.one_hot(torch.Tensor([x]).long(), n_classes).squeeze()
+    return (
+        torch.nn.functional.one_hot(torch.Tensor([x]).long(), n_classes)
+        .squeeze()
+        .float()
+    )
 
 
 def softmax(x: np.ndarray) -> np.ndarray:
